add basic admin ui for setting the shortlink base
The shortlink base URL is set on the general settings page.  I'm using a
bit of javascript to move it up alongside the other URL settings which
results in a slight visual jump, which kinda sucks.

This change also allows defining the shortlink base using the constant
'HUM_SHORTLINK_BASE'.