add support for legacy short url schemes
Allow hum to support the short URL schemes of other plugins or shortlink
generators.  Initially, this just adds support for Friendly Twitter
Links which uses shortlinks of the form /{id} where {id} is the base10
or base32 post ID.  A new 'legacy_hum_id' filter is added to support
additional schemes.  This filter expects a post ID to be returned, which
means this legacy support will only work for WordPress-hosted content.

Resolution of legacy shortlinks happens only after WordPress and all
plugins have failed to resolve the requested URL.  Contrast with
standard Hum shortlinks, which are resolved very early in the process.

Fixes #6